Reorganize photoscan stuff into trk subpackage (#180)
