+++ src/openlifu/db/database.py
@@ -11,8 +11,8 @@ from typing import Dict, List, Union
 
 import h5py
 
-from openlifu.photoscan import Photoscan, load_data_from_photoscan
 from openlifu.plan import Protocol, Run, Solution
+from openlifu.trk.photoscan import Photoscan, load_data_from_photoscan
 from openlifu.util.json import PYFUSEncoder
 from openlifu.xdc import Transducer
