+++ src/openlifu/trk/meshroom_pipelines/default_pipeline.mg
