+++ src/openlifu/trk/meshroom_pipelines/downsample_4x_pipeline.mg
