+++ src/openlifu/trk/meshroom_pipelines/downsample_8x_pipeline.mg
