+++ src/openlifu/trk/meshroom_pipelines/draft_pipeline.mg
